Update virtio_queue.cpp
The constructor can help is delimiting the virtioNet buffer exhaustion by certain. as token is absent the creates a new memory segment by invoking the default constructor, the same could be delimited by creating a new constructor which carries previous buffer token. In that case the previous event will be invoked and new token would be generated . Hence all other events will not be vanished